+++ test/integration/integration/aux/http.clj
@@ -9,3 +9,14 @@
   (let [{:keys [body status]} (test/response-for service-fn :get path)]
     {:status status
      :body   (json/decode body true)}))
+
+(defn post-request-to-test-endpoint
+  [path
+   body
+   service-fn]
+  (let [{:keys [body status]} (test/response-for service-fn
+                                                 :post path
+                                                 :headers {"Content-Type" "application/json"}
+                                                 :body (json/encode body))]
+    {:status status
+     :body   (json/decode body true)}))