+++ test/integration/integration/service_component_test.clj
@@ -1,22 +1,24 @@
 (ns integration.service-component-test
-  (:require [common-clj.component.service :as component.service]
+  (:require [clojure.test :refer :all]
+            [common-clj.component.service :as component.service]
             [common-clj.component.config :as component.config]
             [common-clj.component.routes :as component.routes]
             [com.stuartsierra.component :as component]
-            [clojure.test :refer :all]
             [common-clj.component.helper.core :as component.helper]
-            [integration.aux.http :as aux.http]))
+            [integration.aux.http :as aux.http]
+            [common-clj.error.core :as common-error]))
 
 (def ^:private routes-example [["/test" :get (fn [{{{:keys [datomic-uri]} :config} :components}]
-                                     {:status 200 :body {:test "ok" :datomic-uri datomic-uri}})
-                      :route-name :test]
-                     ["/test-2" :get (fn [_] {:status 200 :body {:test-2 "ok"}}) :route-name :test-2]
-                     ["/throw-info-exception" :get (fn [_] (throw (ex-info "Intentional error"
-                                                                           {:status 42
-                                                                            :cause  :nothing-really})))
-                      :route-name :intentional-info-exception]
-                     ["/throw-simple-exception" :get (fn [_] (throw (throw (Exception. "my exception message"))))
-                      :route-name :intentional-simple-exception]])
+                                               {:status 200 :body {:test "ok" :datomic-uri datomic-uri}})
+                                :route-name :test]
+                               ["/test-2" :get (fn [_] {:status 200 :body {:test-2 "ok"}}) :route-name :test-2]
+                               ["/throw-info-exception" :get (fn [_] (common-error/http-friendly-exception 418
+                                                                                                           "unknown"
+                                                                                                           "I'm not a teapot"
+                                                                                                           "Just chilling"))
+                                :route-name :intentional-info-exception]
+                               ["/throw-simple-exception" :get (fn [_] (throw (throw (Exception. "my exception message"))))
+                                :route-name :intentional-simple-exception]])
 
 (def ^:private system-test
   (component/system-map
@@ -40,8 +42,10 @@
                        :test        "ok"}}
              (aux.http/request-test-endpoints "/test" service-fn))))
     (testing "that we can catch and present friendly error messages using the error interceptor"
-      (is (= {:body   {:cause "nothing-really"}
-              :status 42}
+      (is (= {:status 418
+              :body   {:error   "unknown"
+                       :message "I'm not a teapot"
+                       :detail  "Just chilling"}}
              (aux.http/request-test-endpoints "/throw-info-exception" service-fn)))
       (is (= {:body   {:cause "Internal Server Error"}
               :status 500}